fix: IDENTITY 전략으로 Entity 수정

+++ 07_RestServer/jpa/src/main/java/com/kh/jpa/entity/Board.java
@@ -20,7 +20,7 @@ public class Board {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column
-    private int board_id;
+    private Long board_id;
 
     @Column(length = 100, nullable = false)
     private String board_title;
